requirements.txt
PyQt6==6.8.0
PyQt6-Charts==6.8.0
PyQt6-Charts-Qt6==6.8.1
PyQt6-Qt6==6.8.1
PyQt6-WebEngine==6.8.0
PyQt6-WebEngine-Qt6==6.8.1
PyQt6_sip==13.9.1
python-vlc==3.0.21203
